+++ test/e2e/lib/install.bash
@@ -21,9 +21,11 @@ function uninstall_tiller() {
   kubectl --namespace kube-system delete sa tiller
 }
 
+HELMRELEASE_CRD_URL=https://raw.githubusercontent.com/fluxcd/helm-operator/v1.0.0-rc4/deploy/flux-helm-release-crd.yaml
+
 function install_flux_with_helm() {
 
-  kubectl apply -f https://raw.githubusercontent.com/fluxcd/helm-operator/v1.0.0-rc4/deploy/flux-helm-release-crd.yaml
+  kubectl apply -f "$HELMRELEASE_CRD_URL"
 
   helm install --name flux --wait \
     --namespace "${FLUX_NAMESPACE}" \
@@ -42,7 +44,7 @@ function install_flux_with_helm() {
 
 function uninstall_flux_with_helm() {
   helm delete --purge flux > /dev/null 2>&1
-  kubectl delete crd helmreleases.fluxcd.io > /dev/null 2>&1
+  kubectl delete -f "$HELMRELEASE_CRD_URL" > /dev/null 2>&1
 }
 
 fluxctl_install_cmd="fluxctl install --git-url=ssh://git@gitsrv/git-server/repos/cluster.git --git-email=foo"
@@ -50,7 +52,7 @@ fluxctl_install_cmd="fluxctl install --git-url=ssh://git@gitsrv/git-server/repos
 function install_flux_with_fluxctl() {
   kustomization_dir=${1:-base/flux}
   key_values_varname=${2}
-  kubectl -n "${FLUX_NAMESPACE}" create configmap flux-known-hosts --from-file="${E2E_DIR}/fixtures/known_hosts"
+  kubectl -n "${FLUX_NAMESPACE}" create configmap flux-known-hosts --from-file="${FLUX_ROOT_DIR}/cache/known_hosts"
   local kustomtmp
   kustomtmp="$(mktemp -d)"
   defer rm -rf "'${kustomtmp}'"
@@ -64,6 +66,7 @@ function install_flux_with_fluxctl() {
   if [ -n "$key_values_varname" ]; then
     fill_in_place_recursively "$key_values_varname" "${kustomtmp}"
   fi
+  kubectl apply -f "$HELMRELEASE_CRD_URL"
   kubectl apply -k "${kustomtmp}/${kustomization_dir}" >&3
   kubectl -n "${FLUX_NAMESPACE}" rollout status -w --timeout=30s deployment/flux
 }
@@ -71,6 +74,7 @@ function install_flux_with_fluxctl() {
 function uninstall_flux_with_fluxctl() {
   kubectl delete -n "${FLUX_NAMESPACE}" configmap flux-known-hosts
   $fluxctl_install_cmd --namespace "${FLUX_NAMESPACE}" | kubectl delete -f -
+  kubectl delete -f "$HELMRELEASE_CRD_URL" > /dev/null 2>&1
 }
 
 function install_git_srv() {
@@ -83,7 +87,7 @@ function install_git_srv() {
   defer rm -rf "'$gen_dir'"
   kubectl create secret generic flux-git-deploy \
     --namespace="${FLUX_NAMESPACE}" \
-    --from-file="${FIXTURES_DIR}/known_hosts" \
+    --from-file="${FLUX_ROOT_DIR}/cache/known_hosts" \
     --from-file="$gen_dir/id_rsa" \
     --from-file=identity="$gen_dir/id_rsa" \
     --from-file="$gen_dir/id_rsa.pub"
@@ -120,5 +124,34 @@ function uninstall_git_srv() {
   local secret_name=${1:-flux-git-deploy}
   # Silence secret deletion errors since the secret can be missing (deleted by uninstalling Flux)
   kubectl delete -n "${FLUX_NAMESPACE}" secret "$secret_name" &> /dev/null
-  kubectl delete -n "${FLUX_NAMESPACE}" -f "${E2E_DIR}/fixtures/gitsrv.yaml"
+  kubectl delete -n "${FLUX_NAMESPACE}" -f "${E2E_DIR}/fixtures/kustom/base/gitsrv/gitsrv.yaml"
+}
+
+function install_registry() {
+  local external_access_result_var=${1}
+  local kustomization_dir=${2:-base/registry}
+
+  kubectl apply -n "${FLUX_NAMESPACE}" -k "${E2E_DIR}/fixtures/kustom/${kustomization_dir}" >&3
+
+  # Wait for the registry to be rolled out
+  kubectl -n "${FLUX_NAMESPACE}" rollout status deployment/registry
+
+  local registry_podname
+  registry_podname=$(kubectl get pod -n "${FLUX_NAMESPACE}" -l name=registry -o jsonpath="{['items'][0].metadata.name}")
+  coproc kubectl port-forward -n "${FLUX_NAMESPACE}" "$registry_podname" :5000
+  local local_port
+  read -r local_port <&"${COPROC[0]}"-
+  # shellcheck disable=SC2001
+  local_port=$(echo "$local_port" | sed 's%.*:\([0-9]*\).*%\1%')
+
+  if [ -n "$external_access_result_var" ]; then
+    # return the registry local port, and the PID of the port-forwarding PID into a variable of choice
+    eval "${external_access_result_var}=('$local_port' '$COPROC_PID')"
+  else
+    kill "${COPROC_PID}" > /dev/null
+  fi
+}
+
+function uninstall_registry() {
+  kubectl delete -n "${FLUX_NAMESPACE}" -f "${E2E_DIR}/fixtures/kustom/base/registry/registry.yaml"
 }