+++ test/e2e/run.bash
@@ -10,7 +10,9 @@ FLUX_ROOT_DIR="$(git rev-parse --show-toplevel)"
 E2E_DIR="${FLUX_ROOT_DIR}/test/e2e"
 CACHE_DIR="${FLUX_ROOT_DIR}/cache/$CURRENT_OS_ARCH"
 
-KIND_VERSION="v0.6.1"
+KIND_VERSION=v0.7.0
+KUBE_VERSION=v1.14.10
+GITSRV_VERSION=v1.0.0
 KIND_CACHE_PATH="${CACHE_DIR}/kind-$KIND_VERSION"
 KIND_CLUSTER_PREFIX=flux-e2e
 BATS_EXTRA_ARGS=""
@@ -18,6 +20,11 @@ BATS_EXTRA_ARGS=""
 # shellcheck disable=SC1090
 source "${E2E_DIR}/lib/defer.bash"
 trap run_deferred EXIT
+
+function download_known_hosts() {
+  mkdir -p "${FLUX_ROOT_DIR}/cache"
+  curl -sL "https://github.com/fluxcd/gitsrv/releases/download/${GITSRV_VERSION}/known_hosts.txt" > "${FLUX_ROOT_DIR}/cache/known_hosts"
+}
 
 function install_kind() {
   if [ ! -f "${KIND_CACHE_PATH}" ]; then
@@ -37,15 +44,21 @@ E2E_KIND_CLUSTER_NUM=${E2E_KIND_CLUSTER_NUM:-1}
 if ! kubectl version > /dev/null 2>&1; then
   install_kind
 
+  # We require GNU Parallel, but some systems come with Tollef's parallel (moreutils)
+  if ! parallel -h | grep -q "GNU Parallel"; then
+    echo "GNU Parallel is not available on your system"
+    exit 1
+  fi
+
   echo '>>> Creating Kind Kubernetes cluster(s)'
   KIND_CONFIG_PREFIX="${HOME}/.kube/kind-config-${KIND_CLUSTER_PREFIX}"
-  seq 1 "${E2E_KIND_CLUSTER_NUM}" | time parallel -- env KUBECONFIG="${KIND_CONFIG_PREFIX}-{}" kind create cluster --name "${KIND_CLUSTER_PREFIX}-{}" --wait 5m
   for I in $(seq 1 "${E2E_KIND_CLUSTER_NUM}"); do
     defer kind --name "${KIND_CLUSTER_PREFIX}-${I}" delete cluster > /dev/null 2>&1 || true
     defer rm -rf "${KIND_CONFIG_PREFIX}-${I}"
     # Wire tests with the right cluster based on their BATS_JOB_SLOT env variable
     eval export "KUBECONFIG_SLOT_${I}=${KIND_CONFIG_PREFIX}-${I}"
   done
+  seq 1 "${E2E_KIND_CLUSTER_NUM}" | time parallel -- env KUBECONFIG="${KIND_CONFIG_PREFIX}-{}" kind -v 1 create cluster --name "${KIND_CLUSTER_PREFIX}-{}" --wait 5m --image kindest/node:${KUBE_VERSION}
 
   echo '>>> Loading images into the Kind cluster(s)'
   seq 1 "${E2E_KIND_CLUSTER_NUM}" | time parallel -- kind --name "${KIND_CLUSTER_PREFIX}-{}" load docker-image 'docker.io/fluxcd/flux:latest'
@@ -55,6 +68,7 @@ if ! kubectl version > /dev/null 2>&1; then
 fi
 
 echo '>>> Running the tests'
+download_known_hosts
 # Run all tests by default but let users specify which ones to run, e.g. with E2E_TESTS='11_*' make e2e
 E2E_TESTS=${E2E_TESTS:-.}
 (